requirement.txt
pillow
re
shutil
hachoir